Relocate `noResult` and update imports
Move `noResult` to align with handler module. Update import paths in handler and noResult accordingly for consistency and better organization.
