+++ src/commands/commit/handler.ts
@@ -3,7 +3,7 @@ import { COMMIT_PROMPT } from './prompt'
 import { getApiKeyForModel, getModelAndProviderFromConfig } from '../../lib/langchain/utils'
 import { getLlm } from '../../lib/langchain/utils/getLlm'
 import { getPrompt } from '../../lib/langchain/utils/getPrompt'
-import { noResult } from '../../lib/parsers/noResult'
+import { noResult } from './noResult'
 import { getChanges } from '../../lib/simple-git/getChanges'
 import { CommitArgv, CommitOptions } from './options'
 import { loadConfig } from '../../lib/config/utils/loadConfig'
